+++ IDE/GCC-ARM/include.am
@@ -4,6 +4,7 @@
 
 EXTRA_DIST+= IDE/GCC-ARM/Header/user_settings.h
 EXTRA_DIST+= IDE/GCC-ARM/Source/armtarget.c
+EXTRA_DIST+= IDE/GCC-ARM/Source/wolf_main.c
 EXTRA_DIST+= IDE/GCC-ARM/Source/benchmark_main.c
 EXTRA_DIST+= IDE/GCC-ARM/Source/test_main.c
 EXTRA_DIST+= IDE/GCC-ARM/Source/tls_client.c
